requirements.txt
aiofiles
aiohttp
APScheduler==3.7.0
bs4
emoji==1.6.0
ffmpeg-python
geezlibs==1.6.47
git+https://github.com/Itz-fork/py-trans.git
gitpython
googlesearch-python
gpytranslate
gTTs
gunicorn
heroku3
humanize
httpx
httpx[http2]
motor==3.0.0
opencv-python-headless
prettytable
PyGithub
pymongo
pykeyboard
pymediainfo
python-dotenv
pytgcalls==3.0.0.dev24
pyrogram==2.0.38
reportlab
requests==2.28.2
speedtest-cli==2.1.3
SpeechRecognition
svglib
tgcrypto
telegraph
uwuify
uvloop
wget
youtube-search
youtube-search-python
yt-dlp
